Move python files in tests/ to python/integration. Moved docker/integration.sh to integration/integration.sh
