+++ integration/integration_test.sh
@@ -53,17 +53,17 @@ sleep 5
 
 cat << EOF | parallel --no-notice -n2 -j2 run
 End2End
-tests/end2end_test.py -l ERROR
+python/integration/end2end_test.py -l ERROR
 C2S_extn
-tests/cli_srv_ext_test.py -l ERROR
+python/integration/cli_srv_ext_test.py -l ERROR
 SCMP error
-tests/scmp_error_test.py -l ERROR --runs 60
+python/integration/scmp_error_test.py -l ERROR --runs 60
 Cert/TRC request
-tests/cert_req_test.py -l ERROR
+python/integration/cert_req_test.py -l ERROR
 EOF
 result=$?
 
-run Revocation "tests/revocation_test.sh\
+run Revocation "integration/revocation_test.sh\
  ${REV_BRS:-as1-11:br1-11-3 as2-26:br2-26-2 as1-14:br1-14-3 as1-16:br1-16-2}"
 result=$((result+$?))
