+++ python/integration/path_store_test.py
